+++ lib/src/service/version_check/noop_version_check_service.dart
@@ -2,12 +2,12 @@ import 'package:impaktfull_architecture/src/model/data/version_check/version_che
 import 'package:impaktfull_architecture/src/model/data/version_check/version_requirements.dart';
 import 'package:impaktfull_architecture/src/service/version_check/version_check_service.dart';
 
-class NoopVersionCheckService extends VersionCheckService {
+class ImpaktfullNoopVersionCheckService extends ImpaktfullVersionCheckService {
   @override
-  Future<VersionRequirements?> getVersionRequirements() async => null;
+  Future<ImpaktfullVersionRequirements?> getVersionRequirements() async => null;
 
   @override
-  Future<void> install(VersionCheckResult result) async {}
+  Future<void> install(ImpaktfullVersionCheckResult result) async {}
 
   @override
   void openStore() {}